Workaround for closed time intervals
clj-time's within? method, when used with 2 arguments (a interval and a
date-time), checks whether a date-time is included in the interval,
excluding the end (start <= x <end). When used with 3 arguments
(3 date-times), compare them by checking if (start <= test <= end).

A slightly less verbose workaround to achieve the same behavior is to
include 1 ms in the end date at the time of interval generation so that
both limits are included in the comparison using 2 arguments.